requirements.txt
Pillow==5.0.0
PyMySQL==0.8.0
selenium==3.8.1
